fix: bread units int -> float64

+++ internal/httpserver/request/addproduct.go
@@ -1,7 +1,7 @@
 package request
 
 type AddProduct struct {
-	ProductID  int64  `json:"id" default:"0"`
-	Name       string `json:"name"`
-	BreadUnits int    `json:"bread-units"`
+	ProductID  int64   `json:"id" default:"0"`
+	Name       string  `json:"name"`
+	BreadUnits float64 `json:"bread-units"`
 }
