+++ internal/models/products/proguct.go
@@ -1,7 +1,7 @@
 package products
 
 type Product struct {
-	ProductID  int64  `json:"id"`
-	Name       string `json:"name"`
-	BreadUnits int    `json:"bread-units"`
+	ProductID  int64   `json:"id"`
+	Name       string  `json:"name"`
+	BreadUnits float64 `json:"bread-units"`
 }
